ai_service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.1",
    "google-genai>=1.52.0",
    "httpx>=0.28.1",
    "langchain>=1.0.8",
    "langchain-google-genai>=3.1.0",
    "langchain-mcp-adapters>=0.1.13",
    "langgraph>=1.0.3",
    "mcp-use[search]>=1.4.1",
    "mcp[cli]>=1.22.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
]

[tool.uv.workspace]
members = [
    "weather",
]
